Remove libvlc RefFrames restriction
There is a limitation hardcoded in jellyfin-androidtv that prevents
direct playing a h264 file with libvlc if it has a number of reference
frames == 1.

This limitation was added 5 years ago (da38c02) and causes files to
be transcoded. But since then, libvlc seems to support files with
reference frames < 2.

Thus, this commit removes that limitation, as from my tests
jellyfin-androidtv directs streams and plays those files without any
issue when libvlc is selected.

Closes #958