Refactor duplicated preferences code (#1341)
* Create abstract base class to move common func into

There's a lot of overlap between both preference stores. Additionally,
long-term it would be good to simply abstract away all the property
handling into one location.

That way each store can handle transactions ...etc. but the boilerplate
is in one place

* Unify common functionality from preference stores

Unifies the common bits, this will make it significantly easier to write
future stores.
It also means the most risky parts (de-serializing enums and dynamic
dispatch) are kept in one place. Making debugging / fixes signficantly
easier.

* Add tests for Preference setter/getters

Adds unit tests for the unified setters and getters.
This replicates the work that would be required to add a new store in
the future too.

* Minor tidy-ups

Minor tidy-ups based on other PR comments
- Don't include duplicate interfaces
- Leave formatting alone

* Revert IPreferenceStore rename

Rename IPreferenceStore to PreferenceStore, to match the other
interfaces in the repo

* Merge IPreferenceStore and BasicPreferenceStore

This merges the two classes into a single abstract class
PreferenceScore

We simplify the inheritance chain, keep the various implementation
details in one place and avoid code churn.

An abstract class has to be used so we can mark internal getT and setT
methods protected, and the enum serialization private. An interface
would require us to mark these methods public, polluting the API.

* Address PR comments

- Rename KeyName -> Key
- Move from method body to statement line

* Refactor enum serialization into impl classes

Since this is an implementation detail not all classes may rely on, we
should shift it into the various stores and only test the bare minimum.

* Switch from Junit to KoTest

Switches PreferenceStore test to use KoTest since JUnit has been
completely dropped from the codebase

* Get enum from string backed store directly where possible

In DisplayPreferencesStore we can directly get this from the store and
check for a null, saving a string allocation.

In SharedPreferenceStore, since we don't have access to the backing
store (which was intentional) we call getString instead. Since this is
the only case where null can appear as an empty string value we will
take the small hit to allocate an empty string. That way other users of
the getString interface don't have to check for null types which is a
store implementation detail.

* Reformat Shared code and tidy up test file

- Reformats the SharedPreferenceStoreCode to use less whitespace
- Tidy up the test file, by moving the test class and enum into their
own private sections rather than inside the test class

* Match FunSpec test style to other tests in project

Rather than use an init block we can simply wrap it in the FunSpec
block. This matches the style of all the other tests and removes some
boilerplate too.